refactor login and signup with google and new BE

+++ src/pages/Logout.jsx
@@ -1,6 +1,14 @@
 import { redirect } from "react-router-dom";
+import { logout } from "../utils/http";
 
-export const action = () => {
-  localStorage.clear();
-  return redirect("/login");
+export const action = async () => {
+  try {
+    const response = await logout();
+    if(response.ok) {
+      localStorage.clear();
+      return redirect("/login");
+    } 
+  } catch (error) {
+    return error;
+  }
 };
